attendance2/requirements.txt
asgiref==3.6.0
Django
django-crispy-forms==2.1
django-environ==0.8.1
django-formtools==2.4
django-jazzmin==2.6.0
django-otp==1.1.6
django-phonenumber-field==6.4.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
firebase-admin==6.5.0
future==0.18.2
httplib2==0.22.0
phonenumbers==8.13.11
Pillow==9.5.0
pypng==0.20220715.0
PyQRCode==1.2.1
pytz==2018.7
qrcode==7.4.2
reportlab==4.2.0
sqlparse==0.4.3
validate-email==1.3